% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApacheLogProcessor.R
\name{read.apache.error.log}
\alias{read.apache.error.log}
\title{Read the apache erro log file and loads it to a data frame.}
\usage{
read.apache.error.log(file, columns = c("datetime", "logLevel", "pid",
  "ip_port", "msg"))
}
\arguments{
\item{file}{path to the error log file}

\item{columns}{which columns should be loaded. Default value is all columns. c("datetime", "logLevel", "pid", "ip_port", "msg")}
}
\value{
a data frame with the error log data
}
\description{
Read the apache erro log file and loads it to a data frame.
}
\examples{

#Loads the path of the erro log
path <- system.file("examples", "error_log.log", package = "ApacheLogProcessor")

#Loads the error log to a data frame
dfELog <- read.apache.error.log(path)

}
\author{
Diogo Silveira Mendonca
}
